requirements.txt
playwright
gspread
google-auth
python-dotenv

